requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.6
django-cors-headers==4.2.0
django-environ==0.11.2
django-rest-authtoken==2.1.4
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.103.2
idna==3.4
inflection==0.5.1
ipython==8.16.1
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.6
requests==2.31.0
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
wcwidth==0.2.8
